remove webhook handling
